requirements.txt
streamlit>=1.29.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
gspread>=5.12.4
oauth2client>=4.1.3
python-dotenv>=1.0.0
requests>=2.31.0
plotly>=5.18.0
scipy>=1.11.4
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0
nodeenv>=1.6.0
openpyxl>=3.1.2
typing-extensions>=4.8.0
google-api-python-client>=2.108.0
